Created a command to run both initialize extend arms and extend retract
