+++ src/main/deploy/robotConfigs/SAM/commandconstants.conf
@@ -29,8 +29,8 @@ ArmExtRet {
     tolerance = 0.25
 }
 ArmRotate {
-    Kp = 0.015
-    Ki = 0
+    Kp = 0.01
+    Ki = 0.005
     Kd = 0
     minOutputToMove = 0
     tolerance = 1
