+++ src/main/java/frc/robot/commands/groupCommands/samArmRotExtRetCommands/DefaultArmRotExt.java
@@ -5,7 +5,7 @@
 package frc.robot.commands.groupCommands.samArmRotExtRetCommands;
 
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
-import frc.robot.commands.samArmExtendRetractCommands.ExtendRetract;
+import frc.robot.commands.samArmExtendRetractCommands.ExtendRetractInternal;
 import frc.robot.commands.samArmRotateCommands.EnableArmBrake;
 import frc.robot.subsystems.samArmExtRet.SamArmExtRetBase;
 import frc.robot.subsystems.samArmRotate.SamArmRotateBase;
@@ -17,7 +17,7 @@ public class DefaultArmRotExt extends ParallelCommandGroup {
   /** Creates a new DefaultArmRotExt. */
   public DefaultArmRotExt(SamArmRotateBase armRotate, SamArmExtRetBase armExtRet) {
 
-    addCommands(new EnableArmBrake(armRotate), new ExtendRetract(armExtRet,
+    addCommands(new EnableArmBrake(armRotate), new ExtendRetractInternal(armExtRet,
         ArmRotationExtensionSingleton.getInstance().getPosition(), false));
   }
 }
