+++ src/main/java/frc/robot/oi/SmartDashboard4905.java
@@ -28,7 +28,7 @@ import frc.robot.commands.groupCommands.samArmRotExtRetCommands.StowPosition;
 import frc.robot.commands.groupCommands.samArmRotExtRetCommands.SubstationPickupPosition;
 import frc.robot.commands.groupCommands.samArmRotExtRetCommands.TopScorePosition;
 import frc.robot.commands.limeLightCommands.ToggleLimelightLED;
-import frc.robot.commands.samArmExtendRetractCommands.ExtendRetract;
+import frc.robot.commands.samArmExtendRetractCommands.ExtendRetractInternal;
 import frc.robot.commands.samArmRotateCommands.EnableArmBrake;
 import frc.robot.commands.samArmRotateCommands.RotateArm;
 import frc.robot.commands.showBotCannon.PressurizeCannon;
@@ -94,7 +94,7 @@ public class SmartDashboard4905 {
           new EnableArmBrake(subsystemsContainer.getArmRotateBase()));
       SmartDashboard.putNumber("Extend Arm Position Value", 0);
       SmartDashboard.putData("Extend Arms",
-          new ExtendRetract(subsystemsContainer.getArmExtRetBase(), () -> 5, true, true, false));
+          new ExtendRetractInternal(subsystemsContainer.getArmExtRetBase(), () -> 5, true, true, false));
     SmartDashboard.putData("Arm Angle Tuner", new RotateArm(subsystemsContainer.getArmRotateBase(), ()-> 0, false, true));
     }
